Updated to provide xsi:type for date element in expected data.
This allows different but "equivalent" behavior from IBM DFDL and Daffodil
where 2020-04-02+00:00 is equivalent to 2020-04-02Z.

This also depends on version 0.1.0 of ibm-dfdl-crosstester which insures
that CRLFs are not carried into XML from the DFDL infoset. When XML is constructed, CRLFs are converted to just LF, consistent with Daffodil.

This makes the test work regardless of the fact that the quasiXML data file has
CRLF line endings in it.